Move the machine actuator under a new package
Use a new package for the machine actuator so that we can group all the
actuators based on their topology, rather than having long, weird, file
names.
