+++ pkg/cloud/openstack/machine/actuator.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package openstack
+package machine
 
 import (
 	"context"
@@ -36,8 +36,9 @@ import (
 	bootstrap "sigs.k8s.io/cluster-api-provider-openstack/pkg/bootstrap"
 
 	openstackconfigv1 "sigs.k8s.io/cluster-api-provider-openstack/pkg/apis/openstackproviderconfig/v1alpha1"
+	"sigs.k8s.io/cluster-api-provider-openstack/pkg/cloud/openstack"
 	"sigs.k8s.io/cluster-api-provider-openstack/pkg/cloud/openstack/clients"
-	"sigs.k8s.io/cluster-api-provider-openstack/pkg/cloud/openstack/machinesetup"
+	"sigs.k8s.io/cluster-api-provider-openstack/pkg/cloud/openstack/machine/machinesetup"
 	clusterv1 "sigs.k8s.io/cluster-api/pkg/apis/cluster/v1alpha1"
 	apierrors "sigs.k8s.io/cluster-api/pkg/errors"
 	"sigs.k8s.io/cluster-api/pkg/util"
@@ -45,13 +46,11 @@ import (
 )
 
 const (
-	MachineSetupConfigPath   = "/etc/machinesetup/machine_setup_configs.yaml"
-	SshPrivateKeyPath        = "/etc/sshkeys/private"
-	SshPublicKeyPath         = "/etc/sshkeys/public"
-	SshKeyUserPath           = "/etc/sshkeys/user"
-	CloudConfigPath          = "/etc/cloud/cloud_config.yaml"
-	OpenstackIPAnnotationKey = "openstack-ip-address"
-	OpenstackIdAnnotationKey = "openstack-resourceId"
+	MachineSetupConfigPath = "/etc/machinesetup/machine_setup_configs.yaml"
+	SshPrivateKeyPath      = "/etc/sshkeys/private"
+	SshPublicKeyPath       = "/etc/sshkeys/public"
+	SshKeyUserPath         = "/etc/sshkeys/user"
+	CloudConfigPath        = "/etc/cloud/cloud_config.yaml"
 
 	TimeoutInstanceCreate       = 5 * time.Minute
 	RetryIntervalInstanceStatus = 10 * time.Second
@@ -71,7 +70,7 @@ type OpenstackClient struct {
 	machineSetupWatcher *machinesetup.ConfigWatch
 	machineService      *clients.InstanceService
 	sshCred             *SshCreds
-	*DeploymentClient
+	*openstack.DeploymentClient
 }
 
 func NewMachineActuator(machineClient client.Client, scheme *runtime.Scheme) (*OpenstackClient, error) {
@@ -137,7 +136,7 @@ func NewMachineActuator(machineClient client.Client, scheme *runtime.Scheme) (*O
 		machineSetupWatcher: setupConfigWatcher,
 		scheme:              scheme,
 		sshCred:             &sshCred,
-		DeploymentClient:    NewDeploymentClient(),
+		DeploymentClient:    openstack.NewDeploymentClient(),
 	}, nil
 }
 
@@ -240,7 +239,7 @@ func (oc *OpenstackClient) Delete(cluster *clusterv1.Cluster, machine *clusterv1
 		return nil
 	}
 
-	id := machine.ObjectMeta.Annotations[OpenstackIdAnnotationKey]
+	id := machine.ObjectMeta.Annotations[openstack.OpenstackIdAnnotationKey]
 	err = oc.machineService.InstanceDelete(id)
 	if err != nil {
 		return oc.handleMachineError(machine, apierrors.DeleteMachine(
@@ -387,13 +386,13 @@ func (oc *OpenstackClient) updateAnnotation(machine *clusterv1.Machine, id strin
 	if machine.ObjectMeta.Annotations == nil {
 		machine.ObjectMeta.Annotations = make(map[string]string)
 	}
-	machine.ObjectMeta.Annotations[OpenstackIdAnnotationKey] = id
+	machine.ObjectMeta.Annotations[openstack.OpenstackIdAnnotationKey] = id
 	instance, _ := oc.instanceExists(machine)
 	ip, err := getIPFromInstance(instance)
 	if err != nil {
 		return err
 	}
-	machine.ObjectMeta.Annotations[OpenstackIPAnnotationKey] = ip
+	machine.ObjectMeta.Annotations[openstack.OpenstackIPAnnotationKey] = ip
 	if err := oc.client.Update(nil, machine); err != nil {
 		return err
 	}
