+++ pkg/cloud/openstack/machine/machinesetup/config_types.go
